cleaned up the routes and server. Seperated out draft and leage routes from general football routes
