+++ server.js
@@ -15,6 +15,8 @@ const knexLogger  = require('knex-logger');
 const createNewUser = require('./db/dbFunc/createNewUser');
 
 const footballRoutes = require('./routes/football');
+const leagueRoutes = require('./routes/league');
+const draftRoutes = require('./routes/draft');
 const registerRoutes = require('./routes/register');
 const loginRoutes = require('./routes/login');
 
@@ -40,13 +42,21 @@ app.use(express.static('public'));
 app.get('/', (req, res) => {
   res.redirect('/football');
 });
+
+// football routes
 app.get('/football', footballRoutes(knex));
-app.get('/football/league', footballRoutes(knex));
-app.all('/football/league/create', footballRoutes(knex));
-app.get('/football/league/view', footballRoutes(knex));
-app.get('/football/league/:leagueID', footballRoutes(knex));
-app.all('/football/league/:leagueID/draft/create', footballRoutes(knex));
-app.get('/football/league/:leagueID/draft/:draftID', footballRoutes(knex));
+
+// league routes
+app.get('/football/league', leagueRoutes(knex));
+app.all('/football/league/create', leagueRoutes(knex));
+app.get('/football/league/view', leagueRoutes(knex));
+app.get('/football/league/:leagueID', leagueRoutes(knex));
+
+// draft routes
+app.all('/football/league/:leagueID/draft/create', draftRoutes(knex));
+app.get('/football/league/:leagueID/draft/:draftID', draftRoutes(knex));
+
+// account routes
 app.all('/register', registerRoutes(knex));
 app.all('/login', loginRoutes(knex));
